feat: create a DB view for pair update time digest to accelerate refreshing (#89)
Currently `PairsManager` queries DB to refresh pairs it cached in memory EVERY 1 HOUR. This means that after makers updated pairs, they need to wait for up to 1 hour to see our behavior change.

To accelerate the refreshing while avoid unnecessary performance degrade, we create a DB view for the digest of pairs' update timestamps, which `PairsManger` could rely on to tell whether a heavy pair refreshing operation is needed. As the digest query is light, we can do it every 1 minute.

The DB view is materialized, which means it will use the cached data, instead of querying the original tables when get queried, so our DB won't need to repeat the same digests calculation procedure for each digest query. However, we need to refresh a materialized view after any update to the pairs table. To do it, we added a view refreshing query immediately after every update operation to the pairs table.