feat: Do not build Docker images on Drone. Use Github actions. (#340)
* Do not build Docker images on Drone

We self-host Drone. We use it to build docker images only. In most of
cases we don't have Github integration with Drone. You can't know when
and if the image building process is completed. I guess, it's something
we could try to set up, but we already use Gitub actions to test our
code.

Since, there's no strong reason to use Drone, let's drop it.

* Do not bundle .git into Docker image